fix: introduce `create_tools_json()` and share it with chat_completions.rs
